feat: 훈련주기 관리
