+++ airflow/dockerfile.airflow
@@ -1,9 +1,18 @@
 FROM apache/airflow:2.10.3-python3.9
 
+WORKDIR /opt
+
 COPY airflow/requirements.txt requirements.txt
+COPY airflow/airflow.db airflow.db
+COPY bots/ bots/
+COPY loaders/dataloader.py loaders/dataloader.py
+COPY data/dataloaders data/dataloaders
+COPY data/pdf data/pdf
+COPY summaries/ summaries/
+
 RUN pip install --no-cache-dir -r requirements.txt
 
-ENV AIRFLOW__DATABASE__SQL_ALCHEMY_CONN=sqlite:////Users/leeseungjun/airflow/airflow.db
+ENV AIRFLOW__DATABASE__SQL_ALCHEMY_CONN=sqlite://///airflow.db
 ENV AIRFLOW_HOME=/opt/airflow
 
 WORKDIR $AIRFLOW_HOME
