+++ airflow/requirements.txt
@@ -1,3 +1,12 @@
 apache-airflow-providers-mysql
 python-dotenv
 mysql-connector-python
+python-dotenv
+OpenAI
+pymysql
+mlflow
+scikit-learn
+pandas
+langchain
+langchain_community
+langchain_openai
